tests/wpt/web-platform-tests/custom-elements/instantiating-custom-elements/custom-element-constructor-prototype.html

<!DOCTYPE html>
<html>
<head>
<title>Custom element constructor prototype is the prototype object specified in element definition</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="Custom element constructor prototype is the prototype object specified in element definition">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    var doc = newHTMLDocument();
    var proto = Object.create(SVGElement.prototype);
    var GeneratedConstructor = doc.registerElement('x-a', {
        prototype: proto,
        extends: 'p'
    });
    assert_true(GeneratedConstructor.prototype === proto,
        'Custom element constructor must have the prototype specified in registerElement()');
}, 'If custom element type is registered with prototype, the custom element ' +
   'constructor should have the prototype specified in registerElement() call');
</script>
</body>
</html>
